 \docType{data}
    \name{varambally}
    \alias{varambally}
    \title{The Varambally Cohort from U. Michigan et al. (2005)}
    \format{a \code{ExpressionSet} instance, 1 row per microarray probe.}
    \source{
      GSE3325
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO
    }
        \value{ExpressionSet object for the dataset}
    \author{
Mark Dunning
    }
    \usage{varambally}
    \keyword{datasets}